make roots_down_to_outcome a method of the graph
